For #17: implement WithFileLists without breaking Rpm
